+++ src/components/home/todo/priority-combobox.tsx
@@ -16,7 +16,7 @@ import {
 import { Circle } from "lucide-react";
 import { Todo } from "@/lib/todo_data";
 import { editTodo } from "@/lib/actions";
-import IconLabel from "./icon-label";
+import IconLabel from "../icon-label";
 
 type Priority = {
   value: Todo["priority"];
